requirements.txt
asgiref==3.7.2
Django==4.2.5
django-dotenv==1.4.2
python-dotenv==1.0.1
django-jazzmin==2.6.0
django-querycount==0.8.3
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
psycopg-binary==3.1.12
psycopg2-binary==2.9.8
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
